wip(org): Work on the homepage
