+++ sites/shared/i18n/signup/en.yaml
@@ -19,4 +19,4 @@ err2: Unfortunately, we cannot recover from this error, we need a human being to
 err3: Feel free to try again, or reach out to support so we can assist you.
 signupAgain: Sign up again
 signupLinkExpired: Signup link expired
-
+createAFreeSewingAccount: Create a FreeSewing account
